requirements.txt
PySide6-Fluent-Widgets
PySideSix-Frameless-Window
requests
psutil
py7zr
rarfile
cloudscraper
lxml
openpyxl
pywin32